Move `read_visium_mtx` function to `read_st_data`
Move the  `read_visium_mtx` python function into the `read_st_data`
script, which is the only location where it is currently used.
